requirements.txt
# Core dependencies
pyyaml>=6.0.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-text-splitters>=1.0.0

# Document processing
pypdf>=6.0.0
python-docx>=1.2.0
python-pptx>=0.6.21
unstructured>=0.18.0
markdown>=3.9.0

# Embeddings and vector store
sentence-transformers>=5.0.0
torch>=2.9.0
numpy>=2.3.0
faiss-cpu>=1.12.0
scikit-learn>=1.5.0

# LLM providers
openai>=2.6.0  # For text generation and GPT-4V
mistralai>=1.0.0  # For Document OCR API

# Web framework
fastapi>=0.120.0
uvicorn>=0.38.0
jinja2>=3.1.0

# Utilities
tqdm>=4.67.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Supabase (database + storage + pgvector)
supabase>=2.0.0
vecs>=0.4.0

# Image processing
Pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# Document conversion (for DOCX → PDF)
python-docx-template>=0.16.0  # Fallback if LibreOffice not available
